--- a/package.json
+++ b/package.json
@@ -49,6 +49,7 @@
     "@types/lolex": "^2.1.2",
     "@types/mocha": "^5.2.5",
     "@types/sinon": "^5.0.1",
+    "benchmark": "^2.1.4",
     "chai": "^4.1.2",
     "chai-as-promised": "^7.1.1",
     "codecov": "^3.0.2",