--- a/package.json
+++ b/package.json
@@ -37,9 +37,9 @@
     "bignumber.js": "^7.2.1",
     "debug": "^4.0.0",
     "ilp-logger": "^1.1.3",
-    "ilp-packet": "^3.0.2",
+    "ilp-packet": "^3.0.7",
     "ilp-protocol-ildcp": "^2.0.0",
-    "oer-utils": "^3.2.0",
+    "oer-utils": "^4.0.0",
     "source-map-support": "^0.5.6"
   },
   "devDependencies": {