--- a/src/connection.ts
+++ b/src/connection.ts
@@ -267,6 +267,7 @@ export class Connection extends EventEmitter {
     await this.sendConnectionClose()
     this.safeEmit('end')
     this.safeEmit('close')
+    clearTimeout(this.idleTimer)
   }
 
   /**
@@ -292,6 +293,7 @@ export class Connection extends EventEmitter {
     // wait for all the streams to be closed before emitting the connection 'close'
     await Promise.all(streamClosePromises)
     this.safeEmit('close')
+    clearTimeout(this.idleTimer)
   }
 
   /**