Depends on https://github.com/interledgerjs/oer-utils/pull/21.

Benchmarks (before and after):

```
serialize:   MoneyFrame          v0 x 25,560 ops/sec ±0.84% (93 runs sampled)
serialize:   MoneyFrame          v1 x 116,407 ops/sec ±1.60% (95 runs sampled)
deserialize: MoneyFrame          v0 x 114,379 ops/sec ±2.44% (89 runs sampled)
deserialize: MoneyFrame          v1 x 206,280 ops/sec ±0.55% (90 runs sampled)
```

Also includes misc fix: clean up the idle timer when a connection closes.